requirements.txt
smolagents>=1.24.0
litellm>=1.50.0
polars>=0.20.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
python-dotenv>=1.0.0
rich>=13.0.0
